tools/requirements.txt
# Legacy OpenAI API (for backward compatibility)
openai>=1.0.0

# Google Gemini API (Primary)
google-generativeai>=0.8.0

# LangChain for orchestration
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# Vector Store
chromadb>=0.4.0

# Embeddings (Local)
sentence-transformers>=2.2.0

# Japanese text processing
budoux>=0.7.0

# Retry logic
tenacity>=8.2.0

# Data validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Database
supabase>=2.3.0

